# swagger/src/swagger.json
{
	"components": {
...
		"schemas": {
...
			"CountResponse": {
				"description": "Response for general count",
				"properties": {
					"count": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"count"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"SopType": {
				"description": "SOP type",
				"enum": [
					"General",
					"Special"
				],
				"type": "string"
			},
			"SopResponse": {
				"description": "response of sop apis",
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/SopType"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					},
					"checklists": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"type",
					"startDate",
					"endDate",
					"checklists",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"EntityOverallData": {
				"description": "It stores id and name. Used to minimize entity.",
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"SiteResponse": {
				"description": "response of site apis",
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"noCameras": {
						"type": "number",
						"format": "double"
					},
					"deleted": {
						"type": "boolean"
					},
					"cameras": {
						"items": {
							"$ref": "#/components/schemas/EntityOverallData"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"description",
					"noCameras"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FindAndCountResponse_SiteResponse_": {
				"description": "Response with data array, count, limit and offset. Suited for Paging",
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/SiteResponse"
						},
						"type": "array"
					},
					"count": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"offset": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"count"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SiteRequestBody": {
				"description": "request body of site",
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"cameras": {
						"items": {
							"$ref": "#/components/schemas/EntityOverallData"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SopList": {
				"description": "sop list as response",
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/SopResponse"
						},
						"type": "array"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SopCreateBody": {
				"description": "response of creating sop",
				"properties": {
					"name": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/SopType"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					},
					"checklists": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"type",
					"startDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SopUpdateBody": {
				"description": "request body of updating sop",
				"properties": {
					"name": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/SopType"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"name",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"Authorization": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "Bearer Token get from Login API (JWT)"
			}
		}
	},
	"info": {
		"title": "swagger-service",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "VPS"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/site/list": {
			"get": {
				"operationId": "ListSite",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FindAndCountResponse_SiteResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Site"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/site/count": {
			"get": {
				"operationId": "CountSite",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CountResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Site"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "searchText",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/site/{id}": {
			"get": {
				"operationId": "GetSiteById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SiteResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Site"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteSiteById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SiteResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Site"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateSite",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SiteResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Site"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SiteRequestBody"
							}
						}
					}
				}
			}
		},
		"/site": {
			"post": {
				"operationId": "CreateSite",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SiteResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Site"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SiteRequestBody"
							}
						}
					}
				}
			}
		},
		"/sop/list": {
			"get": {
				"operationId": "ListSop",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SopList"
								}
							}
						}
					}
				},
				"tags": [
					"SOP"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "searchText",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/SopType"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/sop/count": {
			"get": {
				"operationId": "CountSop",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CountResponse"
								}
							}
						}
					}
				},
				"tags": [
					"SOP"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": []
			}
		},
		"/sop/{id}": {
			"get": {
				"operationId": "GetSopById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SopResponse"
								}
							}
						}
					}
				},
				"tags": [
					"SOP"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteSopById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SopResponse"
								}
							}
						}
					}
				},
				"tags": [
					"SOP"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateSop",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SopResponse"
								}
							}
						}
					}
				},
				"tags": [
					"SOP"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SopUpdateBody"
							}
						}
					}
				}
			}
		},
		"/sop": {
			"post": {
				"operationId": "CreateSop",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SopResponse"
								}
							}
						}
					}
				},
				"tags": [
					"SOP"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SopCreateBody"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:9000/apis",
			"description": "Local environment"
		}
	]
}